updates project to use fractal-fork
This updates the project to use fractal-fork instead of the
official fractal image, because the official fractal image
has had a major vulnerability for >1 year. It isn't a big change,
so it should be possible to migrate back to the official fractal
version if a fix is released in the future.

The entire list of changes that differ from the offical image can
be read here:
https://github.com/joyheron/fractal#migration-guide-from-frctlfractal-to-fractal-fork

TL;DR different import structure, much less powerful/colorful CLI, no automatic port
 discovery, fewer dependencies, no vulnerabilities